Fix EmbeddingShard erroneously masking out positional embeddings
Actually I just pasted in a brand new implementation of EmbeddingShard's
__call__ method because the old one is kind of hard to extend.

When EmbeddingShard "embeds" tokens, if you're using a model that uses
fixed positional embeddings (GPT-Neo does, GPT-J doesn't), there is an
additional step required which is to (element-wise) add a constant 2D
array to the 2D array that EmbeddingShard outputs.

The old custom EmbeddingShard would omit this step for tokens with a
token ID larger than or equal to (n_vocab + n_vocab_padding).